update p1 navigation code

+++ value-based method/p1_navigation/dqn_agent.py
@@ -71,9 +71,9 @@ class Agent():
 
         # Epsilon-greedy action selection
         if random.random() > eps:
-            return np.argmax(action_values.cpu().data.numpy()).astype(np.int32)
+            return np.argmax(action_values.cpu().data.numpy())
         else:
-            return random.choice(np.arange(self.action_size)).astype(np.int32)
+            return random.choice(np.arange(self.action_size))
 
     def learn(self, experiences, gamma):
         """Update value parameters using given batch of experience tuples.
@@ -151,6 +151,7 @@ class ReplayBuffer:
         rewards = torch.from_numpy(np.vstack([e.reward for e in experiences if e is not None])).float().to(device)
         next_states = torch.from_numpy(np.vstack([e.next_state for e in experiences if e is not None])).float().to(device)
         dones = torch.from_numpy(np.vstack([e.done for e in experiences if e is not None]).astype(np.uint8)).float().to(device)
+  
         return (states, actions, rewards, next_states, dones)
 
     def __len__(self):
